Add files via upload
This checks the type of variable when it is implemented with an arithmetic operator. Since the variables in AVS language are assigned a type according to the data assigned to them, only the occurrence of variables along with the constants on the right side of the assignment operator is checked.